Add support for a temporary directory in the file store.
Currently, server restart leaves the file store in a partial state; files that
are semi-written will be considered valid after the server restarts. This means
that the reader may receive incorrect data.

This approach creates all files in a temporary directory, and on completion of
writing the file, moves the file into the permanent directory. On FileStore
instantation, all temporary files are deleted.